Handwritten-Digit-Recognizer-PyTorch-CNN/requirements.txt
torch
torchvision
streamlit
streamlit-drawable-canvas
pillow